pyproject.toml
[project]
name = "aws-big-data"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "MICHAL PIATKOWSKI", email = "mcpiatkowski@gmail.com"},
]
dependencies = [
    "pyspark>=3.5.2",
    "mysql-connector-python>=9.0.0",
    "pandas>=2.2.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
